docs: Added zoom example
