+++ projects/f-examples/extensions/zoom-example/zoom-example.component.scss
@@ -1,4 +1,4 @@
-@use "../flow-common";
+@use "../../flow-common";
 
 ::ng-deep zoom-example {
   @include flow-common.connection;
@@ -6,4 +6,8 @@
 
 .f-node {
   @include flow-common.node;
+}
+
+.toolbar {
+  @include flow-common.toolbar;
 }
