+++ projects/f-examples/extensions/zoom-example/zoom-example.component.ts
@@ -1,0 +1,47 @@
+import { ChangeDetectionStrategy, ChangeDetectorRef, Component, ViewChild } from '@angular/core';
+import { FCanvasComponent, FCreateConnectionEvent, FFlowModule, FZoomDirective } from '@foblex/flow';
+import { FCheckboxComponent } from '@foblex/f-docs';
+
+@Component({
+  selector: 'zoom-example',
+  styleUrls: [ './zoom-example.component.scss' ],
+  templateUrl: './zoom-example.component.html',
+  changeDetection: ChangeDetectionStrategy.OnPush,
+  standalone: true,
+  imports: [
+    FFlowModule,
+    FCheckboxComponent
+  ]
+})
+export class ZoomExampleComponent {
+
+  @ViewChild(FCanvasComponent, { static: true })
+  public fCanvas!: FCanvasComponent;
+
+  @ViewChild(FZoomDirective, { static: true })
+  public fZoom!: FZoomDirective;
+
+  public isZoomEnabled: boolean = true;
+
+  constructor(
+    private changeDetectorRef: ChangeDetectorRef
+  ) {
+  }
+
+  public onLoaded(): void {
+    this.fCanvas.resetScaleAndCenter(false);
+  }
+
+  public onZoomIn(): void {
+    this.fZoom.zoomIn();
+  }
+
+  public onZoomOut(): void {
+    this.fZoom.zoomOut();
+  }
+
+  public onZoomOnMouseWheelChanged(checked: boolean): void {
+    this.isZoomEnabled = checked;
+    this.changeDetectorRef.detectChanges();
+  }
+}
