--- projects/f-examples/zoom-example/zoom-example.component.html
@@ -1,7 +1,0 @@
-<f-flow>
-  <f-canvas fZoom>
-    <f-connection fOutputId="output1" fInputId="input1"></f-connection>
-    <div fNode [fNodePosition]="{ x: 24, y: 24 }" fNodeOutput fOutputId="output1" fOutputConnectableSide="right">Node 1</div>
-    <div fNode [fNodePosition]="{ x: 244, y: 24 }" fNodeInput fInputId="input1" fInputConnectableSide="left">Node 2</div>
-  </f-canvas>
-</f-flow>
