--- projects/f-examples/zoom-example/zoom-example.component.ts
@@ -1,16 +1,0 @@
-import { ChangeDetectionStrategy, Component } from '@angular/core';
-import { FFlowModule } from '@foblex/flow';
-
-@Component({
-  selector: 'zoom-example',
-  styleUrls: [ './zoom-example.component.scss' ],
-  templateUrl: './zoom-example.component.html',
-  changeDetection: ChangeDetectionStrategy.OnPush,
-  standalone: true,
-  imports: [
-    FFlowModule
-  ]
-})
-export class ZoomExampleComponent {
-
-}
